added readme and refactored

+++ src/main/java/com/gv/order/management/service/impl/OrderServiceImpl.java
@@ -2,6 +2,7 @@ package com.gv.order.management.service.impl;
 
 import com.gv.order.management.dto.request.OrderRequestDTO;
 import com.gv.order.management.dto.response.OrderResponseDTO;
+import com.gv.order.management.exception.OrderNotFoundException;
 import com.gv.order.management.mapper.OrderMapper;
 import com.gv.order.management.messaging.event.UserUpdatedEvent;
 import com.gv.order.management.model.Order;
@@ -9,6 +10,7 @@ import com.gv.order.management.repository.OrderRepository;
 import com.gv.order.management.service.OrderService;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
 
@@ -16,30 +18,36 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-@RequiredArgsConstructor
 @Service
+@RequiredArgsConstructor
+@Slf4j
 public class OrderServiceImpl implements OrderService {
 
     private final OrderRepository orderRepository;
-
     private final OrderMapper orderMapper;
 
     @Override
     public List<OrderResponseDTO> getAllOrders() {
-        return orderRepository.findAll().stream()
+        log.info("Fetching all orders");
+        final List<OrderResponseDTO> orders = orderRepository.findAll().stream()
                 .map(orderMapper::toOrderResponseDTO)
                 .collect(Collectors.toList());
+        log.debug("Fetched orders: {}", orders);
+        return orders;
     }
 
     @Override
     public OrderResponseDTO getOrderById(final Long id) {
-        final Order order =
-                orderRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Order not found"));
-        return orderMapper.toOrderResponseDTO(order);
+        log.info("Fetching order with ID: {}", id);
+        return orderRepository.findById(id).map(orderMapper::toOrderResponseDTO).orElseThrow(() -> {
+            log.error("Order not found with ID: {}", id);
+            return new OrderNotFoundException(id);
+        });
     }
 
     @Override
     public List<OrderResponseDTO> getOrdersForUser(final Long userId) {
+        log.info("Fetching orders for user ID: {}", userId);
         return orderRepository.findAllByUserId(userId).stream()
                 .map(orderMapper::toOrderResponseDTO)
                 .collect(Collectors.toList());
@@ -47,9 +55,11 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public OrderResponseDTO createOrder(final OrderRequestDTO orderRequestDTO) {
+        log.info("Creating new order with details: {}", orderRequestDTO);
         final Order order = orderMapper.toOrder(orderRequestDTO);
-        final Order savedOrder = orderRepository.save(order);
-        return orderMapper.toOrderResponseDTO(savedOrder);
+        final OrderResponseDTO response = orderMapper.toOrderResponseDTO(orderRepository.save(order));
+        log.debug("Created order: {}", response);
+        return response;
     }
 
     @Override
@@ -67,7 +77,13 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void deleteOrder(final Long id) {
+        log.info("Deleting order with ID: {}", id);
+        if (!orderRepository.existsById(id)) {
+            log.error("Order not found with ID: {}", id);
+            throw new OrderNotFoundException(id);
+        }
         orderRepository.deleteById(id);
+        log.debug("Deleted order with ID: {}", id);
     }
 
     @CacheEvict(value = "orders", allEntries = true)
